+++ config.ini
@@ -8,4 +8,9 @@ memory = memory.json
 temperature = 0.85
 max_tokens = 120
 
-[lilith_img]
+[lilith_display]
+revert_delay = 5
+display_asset_dir = assets
+blink_min_interval = 2
+blink_max_interval = 4
+blink_duration = 0.25
